Add SDK Walkthrough for C++, Move API Funcs, update dummy_c
Add sections that walk through the Go and C++ SDKs. The C++ SDK
walkthrough is complete, the Go SDK Walkthrough needs to be filled in.

Move the individual descriptions of the API Functions to its own page
now that the SDKs do most of the work.

Update the dummy_c walkthrough to reflect its use of the SDK.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>