refactor(docs/plugins): move Go SDK walkthrough in its own section and remove C++ SDK walkthrough
Moving the SDK walkthrough in an ad-hoc section allow more in-depth overviews, and will work better for the future.
The choice of removing the C++ SDK walkthrough is motivated by the fact that it has lately been updated actively, which
makes it not-operational for the most recent plugin API changes. For now, the section has been removed and will be
restored in the future once the C++ SDK is refactored to support the newest API changes.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>