refactor(content/en/docs): remove k8saudit section from the list of supported fields
Since k8saudit became a plugin, it is not part of the default field list supported by Falco and libs.
Moreover, listing k8saudit fields in this section creates a degree of inconsistency towards the practices
used to list the fields of other plugins, which currently provide that kind of documentation inside their
source code in a README file. We can consider having a section dedicated to each plugin on the website too
in the future.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>