The original example results in the following warning:
------
- macro: package_mgmt_binaries
^
------
LOAD_UNUSED_MACRO (Unused macro): Macro not referred to by any other rule/macro

The 'write_binary_dir' rule's condition refers to 'package_mgmt_binaries' as it would be a list of strings when it is actually a macro. The solution is to refer to it as a macro, which fixes the warning and also enables the intended behavior.

Signed-off-by: Tero Kauppinen <tero.kauppinen@est.tech>